server: implement inert resources
When client sends request with new_id argument to the object
that has been destroyed on server-side just before that (client
may not know about it, because it has not got the information
via wire yet), we have to create the resource. If we wouldn't do it then
client will get invalid error id once it tries to use the new object.
But if we create it, then we have to take care that all the requests
but destructor are ignored, because we do not have the server-side
object anymore. Client will destroy the resource right away, because
it gets the information about server-side object destruction.

This patch solves this ugly race by adding wl_resource_set_inert()
function that marks the resource as inert. When resource is inert
it ignores all requests and events but destructor. The trick is in
adding flag into the request/even siganture that says: "hey! I'm
destructor". Server then can skip non-destructor actions on inert
resource.

When an object asks for creating new object from inert object,
wayland will silently create zombie object without invoking
appropriate handler, so once programmer marks resource as inert,
wayland will take care of the rest.

Programmer then can mark newly created resource as inert when
this race come up instead of defining new implementation of
resource just for this rare case.

v2: s/intact/inert
    requests with new_id on inert objects create zombie objects

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>